feat: allow decimal point grades

+++ src/core/application/dtos/monthly-assignments/UpdateMonthlyAssignmentGradeInput.ts
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 export const UpdateMonthlyAssignmentGradeDTO = z.object({
-  grade: z.number().int().min(0).max(100),
+  grade: z.number().min(0).max(100),
 });
 
 export type UpdateMonthlyAssignmentGradeInput = z.infer<typeof UpdateMonthlyAssignmentGradeDTO>;
